backend/requirements.txt
fastapi
uvicorn
python-dotenv
redis
supabase
